UV Index update command (#58)
* uv index model class and generation

* documentation for the uv index table and change to formula lookup variable

* moved UV Index documentation to README.md and running.rst

* basic setup for uv index update command

* uv index update option

* Remove changes to migration readme

+++ woudc_data_registry/product/uv_index/uv_index_generator.py
@@ -59,7 +59,7 @@ from woudc_data_registry.epicentre.metadata import add_metadata
 LOGGER = logging.getLogger(__name__)
 
 
-def execute(path, formula_lookup, bypass):
+def execute(path, formula_lookup, update, start_year, end_year, bypass):
     """
     Orchestrate uv-index generation process
     """
@@ -73,14 +73,33 @@ def execute(path, formula_lookup, bypass):
 
     registry_ = registry.Registry()
 
-    LOGGER.info('erasing current uv index')
+    if not update:
+        LOGGER.info('erasing current uv index')
 
-    registry_.session.query(UVIndex).delete()
-    registry_.save()
+        registry_.session.query(UVIndex).delete()
+        registry_.save()
 
     # traverse directory of files
     for dataset in datasets:
         for dirname, dirnames, filenames in os.walk(dataset):
+            # only ingest years within range for update command
+            if update:
+                split_dir = dirname.split('/')
+                # determine if file year is valid for ingest
+                valid_year = True
+                for base in split_dir:
+                    if base.isnumeric():
+                        int_dir = int(base)
+                        if end_year and int(end_year) < int_dir:
+                            valid_year = False
+                            break
+                        elif start_year and int(start_year) > int_dir:
+                            valid_year = False
+                            break
+
+                if not valid_year:
+                    continue
+
             for filename in filenames:
 
                 ipath = os.path.join(dirname, filename)
@@ -508,7 +527,7 @@ def qa(country, uv):
         return 'E'
 
 
-def generate_uv_index(archivedir, bypass):
+def generate_uv_index(archivedir, update, start_year, end_year, bypass):
     if archivedir is None:
         raise RuntimeError('Missing required on disk archive')
 
@@ -557,5 +576,5 @@ def generate_uv_index(archivedir, bypass):
     LOGGER.info('Loaded formula lookup resource.')
 
     LOGGER.info('Computing UV-index...')
-    execute(archivedir, formula_lookup, bypass)
+    execute(archivedir, formula_lookup, update, start_year, end_year, bypass)
     LOGGER.info('Done.')
